Add brightness and opacity settings
Added a configuration option and UI sliders to adjust what color is applied to the image cover.
The default color with 0.5 brightness and 0.8 opacity was a little too visible in my opinion. Some of the info could be hard to read, and having the play buttons disabled if a requirement was missing looked especially odd. Yes I could have just changed the color value to look better, but, I love customizability so I've added settings.
Very much up to @vcmikuu if this change seems useful.