Fix spacing in UI
Removed the panel-bottom background element and added an anchor-pos-y offset to move everything back in place.
The backgrounds, as far as I can tell, don't load properly and therefore serve no purpose other than to mess up alignment.
This commit is out of the scope of the original PR, please comment if I've introduced a new bug that I can't see.